reuse code for Huge PageRank as well
