+++ algo/src/main/java/org/neo4j/graphalgo/impl/pagerank/NonWeightedComputeStepFactory.java
@@ -1,8 +1,7 @@
 package org.neo4j.graphalgo.impl.pagerank;
 
-import org.neo4j.graphalgo.api.Degrees;
-import org.neo4j.graphalgo.api.RelationshipIterator;
-import org.neo4j.graphalgo.api.RelationshipWeights;
+import org.neo4j.graphalgo.api.*;
+import org.neo4j.graphalgo.core.utils.paged.AllocationTracker;
 
 public class NonWeightedComputeStepFactory implements ComputeStepFactory {
     public ComputeStep createComputeStep(double dampingFactor, int[] sourceNodeIds, RelationshipIterator relationshipIterator, Degrees degrees, RelationshipWeights relationshipWeights, int partitionCount, int start) {
@@ -12,6 +11,20 @@ public class NonWeightedComputeStepFactory implements ComputeStepFactory {
                 relationshipIterator,
                 degrees,
                 relationshipWeights,
+                partitionCount,
+                start
+        );
+    }
+
+    @Override
+    public HugeNonWeightedComputeStep createHugeComputeStep(double dampingFactor, long[] sourceNodeIds, HugeRelationshipIterator relationshipIterator, HugeDegrees degrees, HugeRelationshipWeights relationshipWeights, AllocationTracker tracker, int partitionCount, long start) {
+        return new HugeNonWeightedComputeStep(
+                dampingFactor,
+                sourceNodeIds,
+                relationshipIterator,
+                degrees,
+                relationshipWeights,
+                tracker,
                 partitionCount,
                 start
         );
